backend/requirements.txt
requests
beautifulsoup4
fastapi
uvicorn
pymupdf
pydantic
python-dotenv
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers